refactored network config/utils
