+++ src/components/Modals/index.js
@@ -1,4 +1,3 @@
 export { NetworkSwitchModal } from './NetworkSwitchModal/NetworkSwitchModal'
 export { ErrorModal } from './ErrorModal/ErrorModal'
 export { ConnectModal } from './ConnectModal/ConnectModal'
-// TODO change other modals to named export and aggregate here
